+++ src/components/HeaderComponent.js
@@ -1,45 +1,168 @@
-import { StyleSheet, View, Image, TouchableOpacity, Dimensions, SafeAreaView } from 'react-native';
-import logo from '../../assets/logo.png';
-import { MaterialIcons } from '@expo/vector-icons';
-import { MaterialCommunityIcons } from '@expo/vector-icons';
-import { Octicons } from '@expo/vector-icons';
+import React from 'react'
+import {
+  StyleSheet,
+  View,
+  Image,
+  TouchableOpacity,
+  Dimensions,
+  SafeAreaView,
+  Text
+} from 'react-native';
+import { MaterialIcons, MaterialCommunityIcons, Octicons } from '@expo/vector-icons'
+import { ModalPopup } from './ModalPopup'
+import logo from '../../assets/appAssets/logo.png'
+import botonX from '../../assets/appAssets/x.png'
+import sobre from '../../assets/appAssets/sobre.png'
 
+const { width } = Dimensions.get('window')
+const { height } = Dimensions.get('window')
 
-const {width} = Dimensions.get('window')
-const {height} = Dimensions.get('window')
-
-export default function HeaderComponent(){
+export default function HeaderComponent() {
+  const [visibleObtener, setVisibleObtener] = React.useState(false)
+  const [visibleAnuncio, setVisibleAnuncio] = React.useState(false)
+  const [visibleStickers, setVisibleStickers] = React.useState(false)
   return (
-      <SafeAreaView style= {styles.header}>
-        <TouchableOpacity>
-         <Image 
-            source={logo}
-            style={styles.logo}
-         />  
-        </TouchableOpacity>
+    <SafeAreaView style={styles.header}>
 
-        <TouchableOpacity>
-        <View style={[styles.coins]}>
-            <MaterialIcons name="attach-money" size={25} color="#63130B"/>
-        </View>  
-        </TouchableOpacity>
-
-        <TouchableOpacity>
-        <View style={[styles.cofre]}>
-            <MaterialCommunityIcons name="treasure-chest" size={35} color="#63130B" />
-        </View> 
-        </TouchableOpacity>
-
-        <TouchableOpacity>
-        <View style={[styles.iconos]}>
-            <Octicons name="three-bars" size={35} color="#63130B" />
+      {/* Ventana Emergente de Obtener Cromos */}
+      <ModalPopup visible={visibleObtener}>
+        <View style={{ alignItems: 'center' }}>
+          <View style={styles.modalHeader}>
+            <TouchableOpacity onPress={() => setVisibleObtener(false)}>
+              <Image
+                source={botonX}
+                style={{ height: 30, width: 30 }}
+              />
+            </TouchableOpacity>
+          </View>
         </View>
+        <View style={{ alignItems: 'center' }}>
+          <Image
+            source={sobre}
+            style={{
+              width: 250,
+              height: 250,
+              resizeMode: 'contain',
+            }}
+          />
+        </View>
+        <TouchableOpacity
+          style={styles.logInButton}
+          onPress={() => {
+            setVisibleAnuncio(true)
+            setVisibleStickers(true)
+          }}
+        >
+          <Text style={{ color: 'white', fontWeight: 'bold' }}>Reclamar</Text>
         </TouchableOpacity>
-      </SafeAreaView>
-  );
+      </ModalPopup>
+
+      {/* Ventana Emergente de drop de Stickers */}
+      <ModalPopup visible={visibleStickers}>
+        <View style={{ alignItems: 'center' }}>
+          <View style={styles.modalHeader}>
+            <TouchableOpacity onPress={() => setVisibleStickers(false)}>
+              <Image
+                source={botonX}
+                style={{ height: 30, width: 30 }}
+              />
+            </TouchableOpacity>
+          </View>
+        </View>
+        <View style={{ alignItems: 'center', flexDirection: 'row', flexWrap: 'wrap' }}>
+          <Image
+            source={sobre}
+            style={{ height: 140, width: 140, resizeMode: 'contain', marginVertical: -15 }}
+          />
+          <Image
+            source={sobre}
+            style={{ height: 140, width: 140, resizeMode: 'contain', marginVertical: -15 }}
+          />
+          <Image
+            source={sobre}
+            style={{ height: 140, width: 140, resizeMode: 'contain', marginVertical: -15 }}
+          />
+          <Image
+            source={sobre}
+            style={{ height: 140, width: 140, resizeMode: 'contain', marginVertical: -15 }}
+          />
+          <Image
+            source={sobre}
+            style={{ height: 140, width: 140, resizeMode: 'contain', marginVertical: -15, marginLeft: 75 }}
+          />
+        </View>
+        <Text
+          style={{
+            marginVertical: 30,
+            fontSize: 20,
+            textAlign: 'center'
+          }}>
+          ¡¡Felicidades has obtenido un Sobre!!
+        </Text>
+      </ModalPopup>
+
+      {/* Ventana Emergente de Anuncio */}
+      <ModalPopup visible={visibleAnuncio}>
+        <View style={{ alignItems: 'center' }}>
+          <View style={styles.modalHeader}>
+            <TouchableOpacity onPress={() => setVisibleAnuncio(false)}>
+              <Image
+                source={botonX}
+                style={{ height: 30, width: 30 }}
+              />
+            </TouchableOpacity>
+          </View>
+        </View>
+        <View style={{ alignItems: 'center' }}>
+          <Image
+            source={require('../../assets/Ads/yummy.jpg')}
+            style={{ height: 175, width: 320, resizeMode: 'contain', marginVertical: 10 }}
+          />
+        </View>
+        <Text
+          style={{
+            marginVertical: 30,
+            fontSize: 20,
+            textAlign: 'center'
+          }}>
+          ¡¡Felicidades has obtenido un Sobre!!
+        </Text>
+      </ModalPopup>
+
+      {/* Header Layout */}
+      <Image
+        source={logo}
+        style={styles.logo}
+      />
+
+      <TouchableOpacity>
+        <View style={[styles.coins]}>
+          <MaterialIcons name="attach-money" size={25} color="#63130B" />
+        </View>
+      </TouchableOpacity>
+
+      <TouchableOpacity onPress={() => setVisibleObtener(true)}>
+        <View style={[styles.cofre]}>
+          <MaterialCommunityIcons name="treasure-chest" size={35} color="#63130B" />
+        </View>
+      </TouchableOpacity>
+
+      <TouchableOpacity>
+        <View style={[styles.iconos]}>
+          <Octicons name="three-bars" size={35} color="#63130B" />
+        </View>
+      </TouchableOpacity>
+    </SafeAreaView>
+  )
 }
 
 const styles = StyleSheet.create({
+  modalHeader: {
+    width: '100%',
+    height: 40,
+    alignItems: 'flex-end',
+    justifyContent: 'center',
+  },
   header: {
     width: width,
     height: height / 10,
@@ -72,8 +195,8 @@ const styles = StyleSheet.create({
     marginBottom: 6
   },
   coins: {
-    width:80,
-    height:25,
+    width: 80,
+    height: 25,
     backgroundColor: "#D9D9D9",
     borderRadius: 18,
     alignItems: 'flex-start',
@@ -81,5 +204,13 @@ const styles = StyleSheet.create({
     marginRight: -40,
     marginTop: 3,
     marginBottom: 8
-},
+  },
+  logInButton: {
+    backgroundColor: '#70ABAF',
+    padding: 20,
+    borderRadius: 120,
+    alignItems: 'center',
+    marginVertical: 30,
+    marginHorizontal: 20
+  },
 });
